+++ tables/table-products/table-products.sql
@@ -1,7 +1,7 @@
 CREATE TABLE products(
 	id INT NOT NULL AUTO_INCREMENT,
 	name_product VARCHAR(255) NOT NULL,
-	price INT NOT NULL,
+	price DECIMAL(5,2) NOT NULL,
 	stock INT NOT NULL,
 	description TEXT,
 	category_id INT,
